+++ PluginTemplate/Plugin.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using Dalamud.Game.ClientState.Conditions;
-using Dalamud.Plugin;
+﻿using Dalamud.Plugin;
 using ECommons;
-using ECommons.DalamudServices;
-using ECommons.Reflection;
 using Ocelot;
 using Ocelot.Chain;
 
@@ -16,31 +12,35 @@ public sealed class Plugin : OcelotPlugin
         get => "PluginTemplate";
     }
 
-    public Config config { get; init; }
+    public Config Config { get; }
 
-    public override IOcelotConfig _config => config;
+    public override IOcelotConfig OcelotConfig
+    {
+        get => Config;
+    }
 
-    public static ChainQueue Chain => ChainManager.Get("PluginTemplate##main");
+    public static ChainQueue Chain
+    {
+        get => ChainManager.Get("PluginTemplate.Chain");
+    }
 
     public Plugin(IDalamudPluginInterface plugin)
         : base(plugin, Module.DalamudReflector)
     {
-        config = plugin.GetPluginConfig() as Config ?? new Config();
+        Config = plugin.GetPluginConfig() as Config ?? new Config();
 
-        I18N.SetDirectory(plugin.AssemblyLocation.Directory?.FullName!);
-        I18N.LoadFromFile("en", "Translations/en.json");
+        SetupLanguage(plugin);
 
         OcelotInitialize();
 
         ChainManager.Initialize();
     }
 
-
-    public override bool ShouldTick() => true;
-
-    public override void Dispose()
+    private void SetupLanguage(IDalamudPluginInterface plugin)
     {
-        base.Dispose();
-        ChainManager.Close();
+        I18N.SetDirectory(plugin.AssemblyLocation.Directory?.FullName!);
+        I18N.LoadAllFromDirectory("en", "Translations/en");
+
+        I18N.SetLanguage("en");
     }
 }