+++ PluginTemplate/Windows/ConfigWindow.cs
@@ -3,8 +3,4 @@
 namespace PluginTemplate.Windows;
 
 [OcelotConfigWindow]
-public class ConfigWindow : OcelotConfigWindow
-{
-    public ConfigWindow(Plugin plugin, Config config)
-        : base(plugin, config) { }
-}
+public class ConfigWindow(Plugin plugin, Config config) : OcelotConfigWindow(plugin, config);